Backport misc changes from the fapi repo.

+++ lib/src/upper/rlc_tm.cc
@@ -140,9 +140,9 @@ int rlc_tm::read_pdu(uint8_t* payload, uint32_t nof_bytes)
   if (ul_queue.try_read(&buf)) {
     pdu_size = buf->N_bytes;
     memcpy(payload, buf->msg, buf->N_bytes);
-    log->debug("%s Complete SDU scheduled for tx. Stack latency: %ld us\n",
+    log->debug("%s Complete SDU scheduled for tx. Stack latency: %" PRIu64 " us\n",
                rrc->get_rb_name(lcid).c_str(),
-               buf->get_latency_us().count());
+               (uint64_t)buf->get_latency_us().count());
     log->info_hex(payload,
                   pdu_size,
                   "%s Tx %s PDU, queue size=%d, bytes=%d",
